+++ module/model-firestore/support/service.firestore.ts
@@ -1,8 +1,8 @@
-import type { CommandService } from '@travetto/command';
+import type { ServiceDescriptor } from '@travetto/cli';
 
 const version = process.env.FIRESTORE_VERSION || 'latest';
 
-export const service: CommandService = {
+export const service: ServiceDescriptor = {
   name: 'firestore',
   version,
   ports: { 7000: 8080 },
