+++ module/model-redis/support/service.redis.ts
@@ -1,8 +1,8 @@
-import type { CommandService } from '@travetto/command';
+import type { ServiceDescriptor } from '@travetto/cli';
 
 const version = process.env.REDIS_VERSION || '7.2';
 
-export const service: CommandService = {
+export const service: ServiceDescriptor = {
   name: 'redis',
   version,
   port: 6379,
