+++ module/scaffold/support/bin/features.ts
@@ -74,9 +74,6 @@ export const FEATURES: Feature[] = [
         context: { modelService: 'SQLModelService', modelConfig: 'SQLModelConfig', modelImport: '@travetto/model-sql' }
       }
     ],
-    addons: [
-      { title: 'Command Support', package: '@travetto/command' }
-    ],
     default: 'MongoDB'
   },
 ];
